Visual improvement & Removed Redundant Code
The beat visuals now look much nicer with an opacity applied to them. This makes them feel less cluttered as the player can somewhat see the background through them.

I updated the music player to mimic a desired progression in song difficulty. The game should last about 10 minutes.

+++ Rhythm Survivors/Assets/Scripts/Utility/BeatTimes/Song.cs
@@ -7,5 +7,5 @@ public class Song : MonoBehaviour
 	public AudioClip audioClip;
 	public string jsonFileName;
 	public int loopCount;
-	public int delayOffset;
+	public float delayOffset;
 }
